local changes copied from NPFChan(working)

+++ inc/archive.php
@@ -16,7 +16,7 @@ class Archive {
             return;
 
         // Check if it is a thread
-        $thread_query = prepare(sprintf("SELECT `thread`, `subject`, `body_nomarkup` FROM ``posts_%s`` WHERE `id` = :id", $board['uri']));
+        $thread_query = prepare(sprintf("SELECT `thread`, `subject`, `body_nomarkup`, `trip` FROM ``posts_%s`` WHERE `id` = :id", $board['uri']));
         $thread_query->bindValue(':id', $thread_id, PDO::PARAM_INT);
         $thread_query->execute() or error(db_error($thread_query));
         $thread_data = $thread_query->fetch(PDO::FETCH_ASSOC);
@@ -52,6 +52,9 @@ class Archive {
                 // Remove Post Form from HTML (First Form)
                 $thread_file_content = preg_replace("/<form name=\"post\"(.*?)<\/form>/i", "", $thread_file_content);
 
+                // Refix archive link that will be wrong
+                $thread_file_content = str_replace(sprintf('href="/' . $config['board_path'] . $config['dir']['archive'] . $config['dir']['archive'], $board['uri']), sprintf('href="/' . $config['board_path'] . $config['dir']['archive'], $board['uri']), $thread_file_content);
+
                 // Remove Form from HTML
                 $thread_file_content = preg_replace("/<form(.*?)>/i", "", $thread_file_content);
                 $thread_file_content = preg_replace("/<\/form>/i", "", $thread_file_content);
@@ -64,6 +67,16 @@ class Archive {
                 // Write altered thread HTML to archive location
                 @file_put_contents($board['dir'] . $config['dir']['archive'] . $config['dir']['res'] . sprintf($config['file_page'], $thread_id), $thread_file_content, LOCK_EX);
             }
+
+
+            // Copy json file to Archive
+            // Read Content of Json file
+            $json_file_content = @file_get_contents($board['dir'] . $config['dir']['res'] . json_scrambler($thread_id, true));
+            // Replace links and posting mode to Archived
+            $json_file_content = str_replace(substr($board['dir'], 0, -1) . '\/' . substr($config['dir']['res'], 0, -1), substr($board['dir'], 0, -1) . '\/' . substr($config['dir']['archive'], 0, -1) . '\/' . substr($config['dir']['res'], 0, -1), $json_file_content);
+            // Write altered thread json to archive location
+            @file_put_contents($board['dir'] . $config['dir']['archive'] . $config['dir']['res'] .  json_scrambler($thread_id, true), $json_file_content, LOCK_EX);
+            
 
             // Copy Images and Files Associated with Thread
             if ($post['files']) {
@@ -82,16 +95,22 @@ class Archive {
         }
 
         // Insert Archive Data in Database
-        $query = prepare(sprintf("INSERT INTO ``archive_%s`` VALUES (:thread_id, :snippet, :time, :files, 0)", $board['uri']));
+        $query = prepare(sprintf("INSERT INTO ``archive_%s`` VALUES (:thread_id, :snippet, :lifetime, :files, 0, 0, 0) ON DUPLICATE KEY UPDATE id=:thread_id", $board['uri']));
         $query->bindValue(':thread_id', $thread_id, PDO::PARAM_INT);
         $query->bindValue(':snippet', $thread_data['snippet'], PDO::PARAM_STR);
-        $query->bindValue(':time', 	(!$config['archive']['lifetime'])?0:strtotime("+".$config['archive']['lifetime']." days"), PDO::PARAM_INT);
+        $query->bindValue(':lifetime', 	time(), PDO::PARAM_INT);
         $query->bindValue(':files', json_encode($file_list));
         $query->execute() or error(db_error($query));
 
 
+        // Check if Thread should be Auto Featured based on OP Trip
+        if(in_array($thread_data['trip'], $config['archive']['auto_feature_trips']))
+            self::featureThread($thread_id);
+        
+
         // Purge Threads that have timed out
-        self::purgeArchive();
+        if(!$config['archive']['cron_job']['purge'])
+            self::purgeArchive();
 
         // Rebuild Archive Index
         self::buildArchiveIndex();
@@ -107,8 +126,14 @@ class Archive {
     static public function purgeArchive() {
         global $config, $board;
 
+        // If archive is set to live forever return
+        if(!$config['archive']['lifetime'])
+            return;
+
         // Delete all static pages and files for archived threads that has timed out
-        $query = query(sprintf("SELECT `id`, `files` FROM ``archive_%s`` WHERE `lifetime` < %d AND `featured` = 0", $board['uri'], time())) or error(db_error());
+        $query = prepare(sprintf("SELECT `id`, `files` FROM ``archive_%s`` WHERE `lifetime` < :lifetime AND `featured` = 0 AND `mod_archived` = 0", $board['uri']));
+        $query->bindValue(':lifetime', strtotime("-" . $config['archive']['lifetime']), PDO::PARAM_INT);
+        $query->execute() or error(db_error($query));
         while($thread = $query->fetch(PDO::FETCH_ASSOC)) {
             // Delete Files
             foreach (json_decode($thread['files']) as $f) {
@@ -118,11 +143,22 @@ class Archive {
 
             // Delete Thread
             @unlink($board['dir'] . $config['dir']['archive'] . $config['dir']['res'] . sprintf($config['file_page'], $thread['id']));
+
+            // Delete Vote Data
+            $del_query = prepare("DELETE FROM ``votes_archive`` WHERE `board` = :board AND `thread_id` = :thread_id");
+            $del_query->bindValue(':board', $board['uri']);
+            $del_query->bindValue(':thread_id', $thread['id'], PDO::PARAM_INT);
+            $del_query->execute() or error(db_error($del_query));
         }
 
         // Delete Archive Entries
-        if($query->rowCount() != 0)
-    		$query = query(sprintf("DELETE FROM  ``archive_%s`` WHERE `lifetime` < %d AND `featured` = 0", $board['uri'], time())) or error(db_error());
+        if($query->rowCount() != 0) {
+    		$query = prepare(sprintf("DELETE FROM  ``archive_%s`` WHERE `lifetime` < :lifetime AND `featured` = 0 AND `mod_archived` = 0", $board['uri'])) or error(db_error());
+            $query->bindValue(':lifetime', strtotime("-" . $config['archive']['lifetime']), PDO::PARAM_INT);
+            $query->execute() or error(db_error($query));
+
+            modLog(sprintf("Purged %d archived threads due to expiration date", $query->rowCount()));
+        }
 
         return $query->rowCount();
     }
@@ -133,14 +169,17 @@ class Archive {
 
 
     // Feature thread and replies
-    static public function featureThread($thread_id) {
-        global $config, $board;
+    static public function featureThread($thread_id, $mod_archive = false) {
+        global $config, $board, $mod;
 
         // If featuring of threads is turned off return
-        if(!$config['feature']['threads'])
+        if(!$mod_archive && !$config['feature']['threads'])
+            return;
+        // If mod archive of threads is turned off return
+        if($mod_archive && !$config['mod_archive']['threads'])
             return;
         
-        $query = query(sprintf("SELECT `id`, `files` FROM ``archive_%s`` WHERE `featured` = 0", $board['uri'])) or error(db_error());
+        $query = query(sprintf("SELECT `files` FROM ``archive_%s`` WHERE `id` = %d AND " . ($mod_archive?"`mod_archived`":"`featured`") . " = 0", $board['uri'], (int)$thread_id)) or error(db_error());
         if(!$thread = $query->fetch(PDO::FETCH_ASSOC))
             error($config['error']['invalidpost']);
         
@@ -148,23 +187,29 @@ class Archive {
         $thread_file_content = @file_get_contents($board['dir'] . $config['dir']['archive'] . $config['dir']['res'] . sprintf($config['file_page'], $thread_id));
         
         // Replace links and posting mode to Archived
-        $thread_file_content = str_replace(sprintf('src="/' . $config['board_path'] . $config['dir']['archive'], $board['uri']), sprintf('src="/' . $config['board_path'] . $config['dir']['featured'], $board['uri']), $thread_file_content);
-        $thread_file_content = str_replace(sprintf('href="/' . $config['board_path'] . $config['dir']['archive'], $board['uri']), sprintf('href="/' . $config['board_path'] . $config['dir']['featured'], $board['uri']), $thread_file_content);
+        $thread_file_content = str_replace(sprintf('src="/' . $config['board_path'] . $config['dir']['archive'], $board['uri']), sprintf('src="/' . $config['board_path'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']), $board['uri']), $thread_file_content);
+        $thread_file_content = str_replace(sprintf('href="/' . $config['board_path'] . $config['dir']['archive'], $board['uri']), sprintf('href="/' . $config['board_path'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']), $board['uri']), $thread_file_content);
         $thread_file_content = str_replace('Archived thread', 'Featured thread', $thread_file_content);
 
         // Write altered thread HTML to archive location
-        @file_put_contents($board['dir'] . $config['dir']['featured'] . $config['dir']['res'] . sprintf($config['file_page'], $thread_id), $thread_file_content, LOCK_EX);
+        @file_put_contents($board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['res'] . sprintf($config['file_page'], $thread_id), $thread_file_content, LOCK_EX);
 
         foreach (json_decode($thread['files']) as $f) {
-            @copy($board['dir'] . $config['dir']['archive'] . $config['dir']['img'] . $f->file, $board['dir'] . $config['dir']['featured'] . $config['dir']['img'] . $f->file);
-            @copy($board['dir'] . $config['dir']['archive'] . $config['dir']['thumb'] . $f->thumb, $board['dir'] . $config['dir']['featured'] . $config['dir']['thumb'] . $f->thumb);
+            @copy($board['dir'] . $config['dir']['archive'] . $config['dir']['img'] . $f->file, $board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['img'] . $f->file);
+            @copy($board['dir'] . $config['dir']['archive'] . $config['dir']['thumb'] . $f->thumb, $board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['thumb'] . $f->thumb);
         }
 
         // Update DB entry
-        query(sprintf("UPDATE ``archive_%s`` SET `featured` = 1 WHERE `id` = %d", $board['uri'], (int)$thread_id)) or error(db_error());
+        query(sprintf("UPDATE ``archive_%s`` SET " . ($mod_archive?"`mod_archived`":"`featured`") . " = 1 WHERE `id` = %d", $board['uri'], (int)$thread_id)) or error(db_error());
+
+        // Add mod log entry
+        modLog(sprintf("Added thread #%d to " . ($mod_archive?"mod archive":"featured threads"), $thread_id));
+
 
         // Rebuild Featured Index
         self::buildFeaturedIndex();
+        // Rebuild Archive Index
+        self::buildArchiveIndex();
 
         return true;
     }
@@ -175,31 +220,36 @@ class Archive {
 
 
 
-    static public function deleteFeatured($thread_id) {
-        global $config, $board;
+    static public function deleteFeatured($thread_id, $mod_archive = false) {
+        global $config, $board, $mod;
 
-        $query = query(sprintf("SELECT `id`, `files`, `lifetime` FROM ``archive_%s`` WHERE `featured` = 1", $board['uri'])) or error(db_error());
+        $query = query(sprintf("SELECT `id`, `files`, `lifetime` FROM ``archive_%s`` WHERE `featured` = 1 OR `mod_archived` = 1", $board['uri'])) or error(db_error());
         if(!$thread = $query->fetch(PDO::FETCH_ASSOC))
             error($config['error']['invalidpost']);
         
         
         // Delete Files
         foreach (json_decode($thread['files']) as $f) {
-            @unlink($board['dir'] . $config['dir']['featured'] . $config['dir']['img'] . $f->file);
-            @unlink($board['dir'] . $config['dir']['featured'] . $config['dir']['img'] . $f->thumb);
+            @unlink($board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['img'] . $f->file);
+            @unlink($board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['img'] . $f->thumb);
         }
 
         // Delete Thread
-        @unlink($board['dir'] . $config['dir']['featured'] . $config['dir']['res'] . sprintf($config['file_page'], $thread_id));
+        @unlink($board['dir'] . ($mod_archive?$config['dir']['mod_archive']:$config['dir']['featured']) . $config['dir']['res'] . sprintf($config['file_page'], $thread_id));
 
         // Delete Entry in DB if it has timed out
-        if($thread['lifetime'] != 0 && $thread['lifetime'] < time())
-            query(sprintf("DELETE FROM ``archive_%s`` WHERE `id` = %d", $board['uri'], (int)$thread_id)) or error(db_error());
+        if($thread['lifetime'] != 0 && $thread['lifetime'] < strtotime("-" . $config['archive']['lifetime']))
+            query(sprintf("DELETE FROM ``archive_%s`` WHERE `id` = %d AND " . ($mod_archive?"`featured`":"`mod_archived`") . " = 0", $board['uri'], (int)$thread_id)) or error(db_error());
         else
-            query(sprintf("UPDATE ``archive_%s`` SET `featured` = 0 WHERE `id` = %d", $board['uri'], (int)$thread_id)) or error(db_error());
+            query(sprintf("UPDATE ``archive_%s`` SET " . ($mod_archive?"`mod_archived`":"`featured`") . " = 0 WHERE `id` = %d", $board['uri'], (int)$thread_id)) or error(db_error());
         
+        // Add mod log entry
+        modLog(sprintf("Deleted thread #%d from " . ($mod_archive?"mod archive":"featured threads"), $thread_id));
+
         // Rebuild Featured Index
         self::buildFeaturedIndex();
+        // Rebuild Archive Index
+        self::buildArchiveIndex();
     }
 
 
@@ -212,7 +262,8 @@ class Archive {
             return;
 
         // Purge Archive
-        self::purgeArchive();
+        if(!$config['archive']['cron_job']['purge'])
+            self::purgeArchive();
 
         // Rebuild Archive Index
         self::buildArchiveIndex();
@@ -230,8 +281,8 @@ class Archive {
         if(!$config['archive']['threads'])
             return;
         
-        $query = query(sprintf("SELECT `id`, `snippet`, `featured` FROM ``archive_%s`` WHERE `lifetime` > %d ORDER BY `lifetime` DESC", $board['uri'], time())) or error(db_error());
-        $archive = $query->fetchAll(PDO::FETCH_ASSOC);
+        // Get archive List
+        $archive = self::getArchiveList();
 
         foreach($archive as &$thread)
             $thread['archived_url'] = $config['dir']['res'] . sprintf($config['file_page'], $thread['id']);
@@ -244,10 +295,10 @@ class Archive {
             'boardlist' => createBoardList(false),
             'title' => $title,
             'subtitle' => "",
-            'boardlist' => createBoardlist(),
             'body' => Element("mod/archive_list.html", array(
                 'config' => $config,
-        		'thread_count' => $query->rowCount(),
+        		'thread_count' => count($archive),
+                'board' => $board,
                 'archive' => $archive
             ))
         ));
@@ -255,15 +306,39 @@ class Archive {
         file_write($config['dir']['home'] . $board['dir'] . $config['dir']['archive'] . $config['file_index'], $archive_page);
     }
 
+
+
+    static public function getArchiveList($featured = false, $mod_archive = false, $order_by_lifetime = false) {
+        global $config, $board;
+
+        $archive = false;
+        if($featured) {
+            $query = query(sprintf("SELECT `id`, `snippet`, `featured`, `mod_archived` FROM ``archive_%s`` WHERE `featured` = 1", $board['uri']) . ($order_by_lifetime?" ORDER BY `lifetime` DESC":" ORDER BY `id` DESC")) or error(db_error());
+            $archive = $query->fetchAll(PDO::FETCH_ASSOC);
+        } else if($mod_archive) {
+            $query = query(sprintf("SELECT `id`, `snippet`, `featured`, `mod_archived` FROM ``archive_%s`` WHERE `mod_archived` = 1", $board['uri']) . ($order_by_lifetime?" ORDER BY `lifetime` DESC":" ORDER BY `id` DESC")) or error(db_error());
+            $archive = $query->fetchAll(PDO::FETCH_ASSOC);
+        } else {
+            $query = prepare(sprintf("SELECT `id`, `snippet`, `featured`, `mod_archived`, `votes`  FROM ``archive_%s`` WHERE `lifetime` > :lifetime", $board['uri']) . ($order_by_lifetime?" ORDER BY `lifetime` DESC":" ORDER BY `id` DESC"));
+            $query->bindValue(':lifetime', strtotime("-" . $config['archive']['lifetime']), PDO::PARAM_INT);
+            $query->execute() or error(db_error());
+            $archive = $query->fetchAll(PDO::FETCH_ASSOC);
+        }
+
+        return $archive;
+    }
+
+
+
     static public function buildFeaturedIndex() {
         global $config, $board;
 
         // If featuring of threads is turned off return
         if(!$config['feature']['threads'])
             return;
-        
-        $query = query(sprintf("SELECT `id`, `snippet`, `featured` FROM ``archive_%s`` WHERE `featured` = 1 ORDER BY `lifetime` DESC", $board['uri'])) or error(db_error());
-        $archive = $query->fetchAll(PDO::FETCH_ASSOC);
+
+        // Get featured archived threads
+        $archive = self::getArchiveList(true);
 
         foreach($archive as &$thread)
             $thread['featured_url'] = $config['dir']['res'] . sprintf($config['file_page'], $thread['id']);
@@ -276,9 +351,9 @@ class Archive {
             'boardlist' => createBoardList(false),
             'title' => $title,
             'subtitle' => "",
-            'boardlist' => createBoardlist(),
             'body' => Element("mod/archive_featured_list.html", array(
                 'config' => $config,
+                'board' => $board,
                 'archive' => $archive
             ))
         ));
@@ -286,6 +361,49 @@ class Archive {
         file_write($config['dir']['home'] . $board['dir'] . $config['dir']['featured'] . $config['file_index'], $archive_page);
     }
 
+
+
+
+
+
+    static public function addVote($board, $thread_id) {
+        global $config;
+
+        // Check if thread exist in archive
+        $query = prepare(sprintf("SELECT COUNT(*) FROM ``archive_%s`` WHERE `id` = %d", $board, $thread_id));
+        $query->execute() or error(db_error($query));
+	    if ($query->fetchColumn(0) == 0)
+            error($config['error']['nonexistant']);
+
+        // Check if ip has voted
+        $query = prepare("SELECT COUNT(*) FROM ``votes_archive`` WHERE `board` = :board AND `thread_id` = :thread_id AND `ip` = :ip");
+	    $query->bindValue(':board', $board, PDO::PARAM_STR);
+	    $query->bindValue(':thread_id', $thread_id, PDO::PARAM_INT);
+		$query->bindValue(':ip', get_ip_hash($_SERVER['REMOTE_ADDR']), PDO::PARAM_STR);
+    	
+        // Error if already voted
+        $query->execute() or error(db_error($query));
+	    if ($query->fetchColumn(0) != 0)
+            error($config['error']['already_voted']);
+        
+        // Increase vote count for thread
+        query(sprintf("UPDATE ``archive_%s`` SET `votes` = `votes`+1 WHERE `id` = %d", $board, (int)$thread_id)) or error(db_error());
+
+        // Add ip to voted db
+        $query = prepare("INSERT INTO ``votes_archive`` VALUES (NULL, :board, :thread_id, :ip)  ON DUPLICATE KEY UPDATE");
+	    $query->bindValue(':board', $board, PDO::PARAM_STR);
+	    $query->bindValue(':thread_id', $thread_id, PDO::PARAM_INT);
+		$query->bindValue(':ip', get_ip_hash($_SERVER['REMOTE_ADDR']), PDO::PARAM_STR);
+        $query->execute() or error(db_error($query));
+
+        // Rebuild Archive Index
+        self::buildArchiveIndex();
+    }
+
+
+
+
 }
 
 ?>
+
